[tests] Fix Backend.ETS tests
At some point along the way these tests broke. It seems that the pruning
timer is only set up once (when the first instance of Backend.ETS is
spawned), and so in the "timeout pruning" test the prune timer never
fires, so the bucket never gets pruned.

This indicates an issue with the backend architecture, and with the
structure of the tests, but for now we can fix this by deleting the ETS
table between test runs, so the timer will be initialized on each run.

This commit also includes some other minor tweaks:
- Change the key used in the pruning test
- Use start_supervised in test setup
- Remove the :ets_table_name option to Backend.ETS, which was never
actually used